MAT 918 and MAT 989: Stand alone library packaging changes.
