+++ src/main/java/liquibase/addNewPackagingColumns.1.1.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+		http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-2.0.xsd">
+
+    <changeSet author="mat_dev_user" id="1" context="prod">
+        <preConditions onFail="MARK_RAN">
+            <not><columnExists tableName="MEASURE_EXPORT" columnName="ELM_JSON"/></not>
+            <not><columnExists tableName="MEASURE_EXPORT" columnName="FHIR_XML"/></not>
+            <not><columnExists tableName="MEASURE_EXPORT" columnName="FHIR_LIBS_JSON"/></not>
+            <not><columnExists tableName="CQL_LIBRARY_EXPORT" columnName="ELM_JSON"/></not>
+            <not><columnExists tableName="CQL_LIBRARY_EXPORT" columnName="FHIR_XML"/></not>
+        </preConditions>
+        <addColumn tableName="MEASURE_EXPORT">
+            <column name="ELM_JSON" type="JSON"/>
+        </addColumn>
+        <addColumn tableName="MEASURE_EXPORT">
+            <column name="FHIR_XML" type="longtext"/>
+        </addColumn>
+        <addColumn tableName="MEASURE_EXPORT">
+            <column name="FHIR_LIBS_JSON" type="JSON"/>
+        </addColumn>
+        <addColumn tableName="CQL_LIBRARY_EXPORT">
+            <column name="ELM_JSON" type="JSON"/>
+        </addColumn>
+        <addColumn tableName="CQL_LIBRARY_EXPORT">
+            <column name="FHIR_XML" type="longtext"/>
+        </addColumn>
+    </changeSet>
+</databaseChangeLog>
