+++ src/main/java/liquibase/changelog.xml
@@ -16,7 +16,8 @@
     <include file="classpath:/liquibase/addMeasureFhirVersionColumn.xml"/>
     <include file="classpath:/liquibase/addLibraryFhirVersionColumn.xml"/>
     <include file="classpath:/liquibase/addLibraryQdmVersionColumnNotRequired.xml"/>
-    <include file="classpath:/liquibase/addFhirGlobalLibraryRows.v7.8.xml"/>
     <include file="classpath:/liquibase/upateMissingFhirVersion.xml"/>
     <include file="classpath:/liquibase/addHarpIdColumn.xml"/>
+    <include file="classpath:/liquibase/addNewPackagingColumns.1.1.xml"/>
+    <include file="classpath:/liquibase/addFhirGlobalLibraryRows.v7.9.xml"/>
 </databaseChangeLog>
