--- src/main/java/mat/server/service/FhirLibraryConversionRemoteCall.java
@@ -1,12 +1,0 @@
-package mat.server.service;
-
-
-import mat.dto.fhirconversion.ConversionResultDto;
-import mat.dto.fhirconversion.ConversionType;
-
-public interface FhirLibraryConversionRemoteCall {
-    ConversionResultDto convert(String libraryId, ConversionType conversionType);
-
-
-    ConversionResultDto validate(String libraryId);
-}
